Move UI components into categorical directories
