+++ src/packages/ui/base/Icon.svelte
