+++ src/packages/ui/base/index.ts
@@ -1,0 +1,6 @@
+export { default as Button } from './Button.svelte';
+export { default as FileUploader } from './FileUploader.svelte';
+export { default as Icon } from './Icon.svelte';
+export { default as IconButton } from './IconButton.svelte';
+export { focusOnMount, default as Input } from './Input.svelte';
+export { default as Text } from './Text.svelte';
