+++ src/packages/ui/groups/ArrangedCollection.svelte
