+++ src/packages/ui/groups/index.ts
@@ -1,2 +1,6 @@
+export { default as Arrangable } from './Arrangable.svelte';
+export { default as ArrangedCollection } from './ArrangedCollection.svelte';
+export { default as Collection } from './Collection.svelte';
+export { default as ListTile } from './ListTile.svelte';
 export { default as SelectionGroup } from './SelectionGroup.svelte';
 export type { SelectionContext, SelectOptions } from './SelectionGroup.svelte';
