+++ src/packages/ui/index.ts
@@ -1,19 +1,5 @@
-export { default as Arrangable } from './Arrangable.svelte';
-export { default as ArrangedCollection } from './ArrangedCollection.svelte';
-export { default as Button } from './Button.svelte';
-export { default as Collection } from './Collection.svelte';
-export { default as Content } from './Content.svelte';
-export { default as Draggable, draggable } from './Draggable.svelte';
-export type { DraggableParams } from './Draggable.svelte';
-export { default as DragHandle } from './DragHandle.svelte';
-export { default as FileUploader } from './FileUploader.svelte';
-export { default as Icon } from './Icon.svelte';
-export { default as IconButton } from './IconButton.svelte';
-export { focusOnMount, default as Input } from './Input.svelte';
-export { default as ListTile } from './ListTile.svelte';
-export { default as PanView } from './PanView.svelte';
-export { default as Text } from './Text.svelte';
-
+export * from './base';
+export * from './dragging';
 export * from './groups';
 export * from './layout';
 export * from './modal';
